requirements.txt
Flask==2.3.3
flask-limiter==3.5.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0